Revert "Use multi release jar to move the Jackson v3 implementation in vertx-core."
This reverts commit ef9f7b21768b70eddb00a0761e063659fba9f8eb.
