--- vertx-core-jackson-v3/src/test/java/module-info.java
@@ -1,8 +1,0 @@
-
-open module io.vertx.core.jacksonv3.tests {
-
-  requires io.vertx.core;
-
-  requires junit;
-
-}
