--- vertx-core/src/main/java25/io/vertx/core/json/jackson/JacksonFactory.java
@@ -1,45 +1,0 @@
-/*
- * Copyright (c) 2011-2019 Contributors to the Eclipse Foundation
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
- * which is available at https://www.apache.org/licenses/LICENSE-2.0.
- *
- * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
- */
-
-package io.vertx.core.json.jackson;
-
-import io.vertx.core.spi.json.JsonCodec;
-
-/**
- * @author <a href="mailto:julien@julienviet.com">Julien Viet</a>
- */
-public class JacksonFactory implements io.vertx.core.spi.JsonFactory {
-
-  public static final JacksonFactory INSTANCE = new JacksonFactory();
-
-  public static final JsonCodec CODEC;
-
-  static {
-    JsonCodec codec;
-    try {
-      codec = new io.vertx.core.json.jackson.v3.JacksonCodec();
-    } catch (Throwable ignore1) {
-      // Not available
-      try {
-        codec = new DatabindCodec();
-      } catch (Throwable ignore2) {
-        // No databind
-        codec = new JacksonCodec();
-      }
-    }
-    CODEC = codec;
-  }
-
-  @Override
-  public JsonCodec codec() {
-    return CODEC;
-  }
-}
